pyproject.toml
[project]
name = "streaming-grpc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "grpcio>=1.74.0",
    "grpcio-tools>=1.74.0",
    "protobuf>=6.32.0",
]
